+++ cpp/src/io/csv/csv_reader.cu
@@ -1573,6 +1573,20 @@ gdf_error launch_dataTypeDetection(raw_csv_t *raw_csv,
   return GDF_SUCCESS;
 }
 
+/**
+* @brief Returns true is the input character is a valid digit.
+* Supports both decimal and hexadecimal digits (uppercase and lowercase).
+*/
+__device__ __forceinline__
+bool isDigit(char c, bool is_hex){
+	if (c >= '0' && c <= '9') return true;
+	if (is_hex) {
+		if (c >= 'A' && c <= 'F') return true;
+		if (c >= 'a' && c <= 'f') return true;
+	}
+	return false;
+}
+
 /**---------------------------------------------------------------------------*
  * @brief CUDA kernel that parses and converts CSV data into cuDF column data.
  *
@@ -1646,10 +1660,13 @@ void dataTypeDetection(char *raw_csv,
 			// This could possibly result in additional empty fields
 			adjustForWhitespaceAndQuotes(raw_csv, &start, &tempPos);
 
-			long strLen=tempPos-start+1;
+			const long strLen = tempPos - start + 1;
+
+			const bool maybe_hex = ((strLen > 2 && raw_csv[start] == '0' && raw_csv[start + 1] == 'x') ||
+				(strLen > 3 && raw_csv[start] == '-' && raw_csv[start + 1] == '0' && raw_csv[start + 2] == 'x'));
 
 			for(long startPos=start; startPos<=tempPos; startPos++){
-				if(raw_csv[startPos]>= '0' && raw_csv[startPos] <= '9'){
+				if(isDigit(raw_csv[startPos], maybe_hex)){
 					countNumber++;
 					continue;
 				}
@@ -1669,11 +1686,21 @@ void dataTypeDetection(char *raw_csv,
 				}
 			}
 
+			// Integers have to have the length of the string
+			long int_req_number_cnt = strLen;
+			// Off by one if they start with a minus sign
+			if(raw_csv[start]=='-' && strLen > 1){
+				--int_req_number_cnt;
+			}
+			// Off by one if they are a hexadecimal number
+			if(maybe_hex) {
+				--int_req_number_cnt;
+			}
+
 			if(strLen==0){ // Removed spaces ' ' in the pre-processing and thus we can have an empty string.
 				atomicAdd(& d_columnData[actual_col].countNULL, 1L);
 			}
-			// Integers have to have the length of the string or can be off by one if they start with a minus sign
-			else if(countNumber==(strLen) || ( strLen>1 && countNumber==(strLen-1) && raw_csv[start]=='-') ){
+			else if(countNumber==int_req_number_cnt){
 				// Checking to see if we the integer value requires 8,16,32,64 bits.
 				// This will allow us to allocate the exact amount of memory.
 				const auto value = convertStrToValue<int64_t>(raw_csv, start, tempPos, opts);