+++ python/cudf/tests/test_csvreader.py
@@ -918,3 +918,20 @@ def test_csv_reader_aligned_byte_range(tmpdir):
     df = cudf.read_csv(str(fname), byte_range=(0, 4096))
     # read_csv call above used to crash; the assert below is not crucial
     assert(np.count_nonzero(df['zeros']) == 0)
+
+
+def test_csv_reader_hex_ints(tmpdir):
+    lines = ['0x0', '-0x1000', '0xfedcba', '0xABCDEF', '0xaBcDeF']
+    values = [int(hex_int, 16) for hex_int in lines]
+
+    buffer = '\n'.join(lines) + '\n'
+
+    # with explicit data types
+    df = read_csv(StringIO(buffer),
+                  dtype=['int32'], names=['hex_int'])
+    np.testing.assert_array_equal(values, df['hex_int'])
+
+    # with data type inference
+    df = read_csv(StringIO(buffer),
+                  names=['hex_int'])
+    np.testing.assert_array_equal(values, df['hex_int'])